refactor(session): cleanup sessions
